pyproject.toml
[project]
name = "rag-demo"
version = "0.1.0"
description = "RAG + agentic coding assistant demo for Phison aiDAPTIV"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.9.0",
]

[project.optional-dependencies]
dev = ["pytest", "httpx", "psutil"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend", "config", "ingest"]

